# GGRC-1493 GGRC-1626 Improve export performance