--- a/src/ggrc/converters/base.py
+++ b/src/ggrc/converters/base.py
@@ -6,6 +6,7 @@
 from collections import defaultdict
 
 from ggrc import settings
+from ggrc.utils import benchmark
 from ggrc.utils import structures
 from ggrc.cache.memcache import MemCache
 from ggrc.converters import get_exportables
@@ -64,9 +65,12 @@ class Converter(object):
     self.indexer = get_indexer()
 
   def to_array(self):
-    self.block_converters_from_ids()
-    self.handle_row_data()
-    return self.to_block_array()
+    with benchmark("Create block converters"):
+      self.block_converters_from_ids()
+    with benchmark("Handle row data"):
+      self.handle_row_data()
+    with benchmark("Make block array"):
+      return self.to_block_array()
 
   def to_block_array(self):
     """ exporting each in it's own block separated by empty lines