--- a/src/ggrc/converters/base_block.py
+++ b/src/ggrc/converters/base_block.py
@@ -98,6 +98,7 @@ class BlockConverter(object):
     self._mapping_cache = None
     self._owners_cache = None
     self._roles_cache = None
+    self._user_roles_cache = None
     self._ca_definitions_cache = None
     self.converter = converter
     self.offset = options.get("offset", 0)
@@ -246,6 +247,40 @@ class BlockConverter(object):
       self._roles_cache = {role.name: role for role in
                            models.all_models.Role.query}
     return self._roles_cache[name]
+
+  def _create_user_roles_cache(self):
+    """Create cache for user roles.
+
+    The cache returns a list of emails for a role in a context.
+    """
+    cache = defaultdict(lambda: defaultdict(set))
+    context_ids = {rc.obj.context_id for rc in self.row_converters}
+    user_roles = db.session.query(
+        models.all_models.UserRole.context_id,
+        models.all_models.UserRole.role_id,
+        models.all_models.UserRole.person_id,
+    ).filter(
+        models.all_models.UserRole.context_id.in_(context_ids)
+    ).all()
+
+    people_ids = {role[2] for role in user_roles}
+
+    emails_map = dict(db.session.query(
+        models.Person.id,
+        models.Person.email
+    ).filter(
+        models.Person.id.in_(people_ids)
+    ))
+
+    for context_id, role_id, person_id in user_roles:
+      cache[context_id][role_id].add(emails_map[person_id])
+    return cache
+
+  def get_user_roles_cache(self):
+    """Get cache for emails on user roles by context."""
+    if self._user_roles_cache is None:
+      self._user_roles_cache = self._create_user_roles_cache()
+    return self._user_roles_cache
 
   def _create_owners_cache(self):
     """Create a cache of emails for all object owners."""