--- a/src/ggrc/converters/handlers/handlers.py
+++ b/src/ggrc/converters/handlers/handlers.py
@@ -280,7 +280,9 @@ class OwnerColumnHandler(UserColumnHandler):
       )
 
   def get_value(self):
-    emails = [owner.email for owner in self.row_converter.obj.owners]
+    """Get a list of object owner emails."""
+    cache = self.row_converter.block_converter.get_owners_cache()
+    emails = [cache[owner.id] for owner in self.row_converter.obj.owners]
     return "\n".join(emails)