--- a/src/ggrc/converters/handlers/handlers.py
+++ b/src/ggrc/converters/handlers/handlers.py
@@ -282,7 +282,8 @@ class OwnerColumnHandler(UserColumnHandler):
   def get_value(self):
     """Get a list of object owner emails."""
     cache = self.row_converter.block_converter.get_owners_cache()
-    emails = [cache[owner.id] for owner in self.row_converter.obj.owners]
+    emails = [cache[owner.person_id]
+              for owner in self.row_converter.obj.object_owners]
     return "\n".join(emails)