--- a/src/ggrc/views/converters.py
+++ b/src/ggrc/views/converters.py
@@ -54,20 +54,24 @@ def parse_export_request():
 
 def handle_export_request():
   try:
-    data = parse_export_request()
-    query_helper = QueryHelper(data)
-    converter = Converter(ids_by_type=query_helper.get_ids())
-    csv_data = converter.to_array()
-    csv_string = generate_csv_string(csv_data)
-
-    object_names = "_".join(converter.get_object_names())
-    filename = "{}.csv".format(object_names)
-
-    headers = [
-        ("Content-Type", "text/csv"),
-        ("Content-Disposition", "attachment; filename='{}'".format(filename)),
-    ]
-    return current_app.make_response((csv_string, 200, headers))
+    with benchmark("handle export request"):
+      data = parse_export_request()
+      query_helper = QueryHelper(data)
+      ids_by_type = query_helper.get_ids()
+    with benchmark("Generate CSV array"):
+      converter = Converter(ids_by_type=ids_by_type)
+      csv_data = converter.to_array()
+    with benchmark("Generate CSV string"):
+      csv_string = generate_csv_string(csv_data)
+    with benchmark("Make response."):
+      object_names = "_".join(converter.get_object_names())
+      filename = "{}.csv".format(object_names)
+      headers = [
+          ("Content-Type", "text/csv"),
+          ("Content-Disposition",
+           "attachment; filename='{}'".format(filename)),
+      ]
+      return current_app.make_response((csv_string, 200, headers))
   except BadQueryException as exception:
     raise BadRequest(exception.message)
   except:  # pylint: disable=bare-except