Exporting 2K objects with a lot of mappings caused app-engine to timeout and throw an error.

This PR fixes currently the biggest bottleneck in exports with mappings.

Benchmarks (mapping cache for 2k regulations and around 100 different mappings):
before: 18.4923s
after: 0.3593s

Additional benchmark (exporting all Programs, Controls and regulations from grc-test db):
before: 169.3381 seconds - 98001 queries
after: 39.4760 seconds - 14502 queries 
after 5f7937b: 38.9554 seconds - 9485 queries
after 0d302d0: 34.9496 seconds - 3433 queries