@Smotko 2017-03-24
We now keep a lot more objects in python memory (cache), can you do a quick check to see how much more memory we now need for export?