@zidarsk8 2017-03-24
@Smotko We only pull data that is going to be included in the CSV file, which means that If that was too much, we would not be able to create and hold the CSV in ram and the export would fail anyway. But from my testing, the caches that only hold short string did not affect memory usage in any significant way.