@zidarsk8 2017-03-24
Before and after for exporting:
Regulations (231),
Controls (2666) - on before the request failed because process ran out of memory,
Objectives (862),
Programs (989),

Before:
![regulation 231 -objective 862 -program 989](https://cloud.githubusercontent.com/assets/566311/24317134/682a33ea-10ec-11e7-8178-6436736e89b3.png)

After:
![screenshot from 2017-03-24 23-46-44](https://cloud.githubusercontent.com/assets/566311/24317138/6cb59242-10ec-11e7-9950-4560429a7525.png)